green: isolate command palette coverage
